requirements.txt
streamlit
python-dotenv
yt-dlp
langchain
langchain-google-genai
langchain-community
langchain-core
faiss-cpu
grpcio
google-api-core
google-auth
protobuf
